% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_viz.R
\name{comparison_viz}
\alias{comparison_viz}
\title{Generating Model Comparison Viz}
\usage{
comparison_viz(comparison, choice = "accuracy")
}
\arguments{
\item{comparison}{A dataframe, the output of train_test_acc_time(). Dimensions are 7 columns and at least 1 row.}

\item{choice}{String, selecting the desired comparison visualization; either "accuracy" or "time". Default value is 'accuracy'.}
}
\value{
PNG image; A bar chart of accuracies or a time comparison bar chart, saved to root directory.
}
\description{
This function visualizes the output of train_test_acc_time() for easy communication and interpretation. The user has the choice to visualize a comparison of accuracies or time.
}
\details{
This function outputs a beautiful ggplot bar chart comparison of different models' training and test scores or the time it takes to fit and predict.
}
\examples{
results_dataframe <- data.frame(model = c('svmPoly','knn','rf'), 
Train_acc = runif(3,0,1), 
Test_acc = runif(3,0,1), 
Variance = runif(3,0,1), 
Fit_Time = runif(3,0,4), 
Predict_Time = runif(3,0,4), 
Total_Time = runif(3,0,4))
comparison_viz(results_dataframe)
comparison_viz(results_dataframe, "time")

}
