% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Splitting data into train, validation, train_validation and test sets}
\usage{
split(X, y, prop_train, prop_test, prop_validation)
}
\arguments{
\item{X}{A dataframe of training data}

\item{y}{A dataframe of the target data}

\item{prop_train}{A float; proportion of data to be split into a training set. Prop_train, prop_test, and prop_validation need to add to 1.}

\item{prop_test}{A float; proportion of data to be split into a test set. Prop_train, prop_test, and prop_validation need to add to 1.}

\item{prop_validation}{A float; proportion of data to be split into a validation set. Prop_train, prop_test, and prop_validation need to add to 1.}
}
\value{
Four X dataframes and four y lists. One each for training, validation, test, and a combined training and validation set.
}
\description{
This function splits the training set X, and target set y (class labels in classification, real numbers in regression) into 
train, validation, train & validation, and test sets according to specified proportions. Proportions must add up to 1.
}
\details{
Outputs four X dataframes and four y lists. One each for training, validation, test, and a combined training and validation set.
}
\examples{
split(X_dataset, y_dataset, 0.4, 0.2, 0.2)
split(X_toy_data, y_toy_data, 0.7, 0.1, 0.2)
}
